requirements.txt
streamlit>=1.31.0
ultralytics>=8.0.0
opencv-python-headless
numpy
pandas
pillow
torch
torchvision
